requirements.txt
fastapi==0.133.0
uvicorn==0.41.0
sqlalchemy==2.0.46
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
requests==2.32.5
python-jose==3.5.0
email-validator==2.3.0
pytest==9.0.2
httpx==0.28.1
